use integer with gl::makeVersion instead of float

+++ libvpvl2/include/vpvl2/IApplicationContext.h
@@ -118,7 +118,7 @@ public:
         virtual ~FunctionResolver() {}
         virtual bool hasExtension(const char *name) const = 0;
         virtual void *resolveSymbol(const char *name) const = 0;
-        virtual float query(QueryType type) const = 0;
+        virtual int query(QueryType type) const = 0;
     };
 
     struct TextureDataBridge {
